refactor: Fix DI issues
