+++ src/auth/auth.controller.ts
@@ -7,12 +7,13 @@ import { Response } from 'express'
 
 @Controller('api/auth')
 export class AuthController {
-    // init
     constructor(private authService: AuthService) { }
 
     // Sign-Up
     @Post('/signup')
-    async singUp(@Body() createUserDTO: CreateUserDTO): Promise<ApiResponseDTO<void>> {
+    async singUp(
+        @Body() createUserDTO: CreateUserDTO
+    ): Promise<ApiResponseDTO<void>> {
         await this.authService.signUp(createUserDTO)
 
         return new ApiResponseDTO(true, HttpStatus.CREATED, 'User Created Successfully!')
@@ -20,7 +21,10 @@ export class AuthController {
 
     // Sign-In
     @Post('/signin')
-    async singIn(@Body() signInDTO: SignInDTO, @Res() res: Response): Promise<void> {
+    async singIn(
+        @Body() signInDTO: SignInDTO,
+        @Res() res: Response
+    ): Promise<void> {
         const accessToken = await this.authService.signIn(signInDTO)
 
         // [2] JWT를 헤더에 저장 후 ApiResponse를 바디에 담아 전송
