+++ src/manager-requests/manager-requests.controller.ts
@@ -1,3 +1,4 @@
+import { GetUser } from './../common/custom-decorators/get-user.decorator';
 import { Body, Controller, Delete, ForbiddenException, Get, HttpStatus, NotFoundException, Param, Post, Put, Req, UseGuards } from '@nestjs/common'
 import { ManagerRequestsService } from './manager-requests.service'
 import { CreateManagerRequestDTO } from './DTO/create-manager-request.dto'
@@ -11,12 +12,11 @@ import { Roles } from 'src/common/custom-decorators/roles.decorator'
 import { AuthenticatedRequest } from 'src/auth/interfaces/authenticated-request.interface'
 
 @Controller('api/manager-requests')
-@UseGuards(AuthGuard('jwt'), RolesGuard) // JWT인증, roles guard 적용
+@UseGuards(AuthGuard('jwt'), RolesGuard)
 export class ManagerRequestsController {
     constructor(private managerRequestsService: ManagerRequestsService) { }
 
     // CREATE
-    // 점주 신청서 생성
     @Post('/')
     @Roles(UserRole.USER)
     async createManagerRequest(
@@ -30,7 +30,6 @@ export class ManagerRequestsController {
     }
 
     // READ
-    // 모든 점주 신청서 조회 (관리자 전용)
     @Get('/')
     @Roles(UserRole.ADMIN)
     async readAllManagerRequests(): Promise<ApiResponseDTO<ReadManagerRequestDTO[]>> {
@@ -44,10 +43,12 @@ export class ManagerRequestsController {
         return new ApiResponseDTO(true, HttpStatus.OK, "Requests Retrieved Successfully", readManagerRequestDTO)
     }
 
-    // 특정 점주 신청서 조회
+    // 특정 점주신청서 조회
     @Get('/:request_id')
     @Roles(UserRole.ADMIN)
-    async readManagerRequestById(@Param('request_id') request_id: number): Promise<ApiResponseDTO<ReadManagerRequestDTO>> {
+    async readManagerRequestById(
+        @Param('request_id') request_id: number
+    ): Promise<ApiResponseDTO<ReadManagerRequestDTO>> {
         const managerRequest = await this.managerRequestsService.readManagerRequestById(request_id)
         if (!managerRequest) {
             throw new NotFoundException(`Cannot Find Request by Id ${request_id}`)
@@ -57,18 +58,50 @@ export class ManagerRequestsController {
         return new ApiResponseDTO(true, HttpStatus.OK, "Request Retrieved Successfully", readManagerRequestDTO)
     }
 
-    // UPDATE
-    // 가게 신청서 처리 (관리자 전용)
+    // 로그인된 회원의(나의) 점주신청서 조회
+    @Get('/my')
+    @Roles(UserRole.USER)
+    async readMyManagerRequest(
+        @Req() req: AuthenticatedRequest
+    ): Promise<ApiResponseDTO<ReadManagerRequestDTO[]>> {
+        const logginedUser = req.user.user_id
+        const managerRequest = await this.managerRequestsService.readManagerRequestByUser(logginedUser)
+        if (!managerRequest) {
+            throw new NotFoundException(`Cannot Find Request by Id ${logginedUser}`)
+        }
+
+        const readManagerRequestDTOs = managerRequest.map(request => new ReadManagerRequestDTO(request));
+        return new ApiResponseDTO(true, HttpStatus.OK, "Request Retrieved Successfully", readManagerRequestDTOs)
+    }
+    
+    // READ - 특정아이디() 점주 신청서 조회 - ///// 바로 위랑 중복되는거인듯? /////
+    // @Get('/my-manager-requests')
+    // @UseGuards(AuthGuard('jwt'), RolesGuard)
+    // @Roles(UserRole.MANAGER)
+    // async readMyManagerRequests(@Req() req: AuthenticatedRequest): Promise<ApiResponseDTO<ReadManagerRequestDTO[]>> {
+    //     const user_id = req.user.user_id
+
+    //     // const foundRequests = await this.managerRequestsService.readManagerRequestByUser(user_id)
+    //     const foundRequests = await this.managerRequestsService.readManagerRequestByUser(user_id)
+    //     const readManagerRequestDTO = foundRequests.map((request) => new ReadManagerRequestDTO(request))
+
+    //     return new ApiResponseDTO(true, HttpStatus.OK, 'My ManagerRequests Retrieved Successfully', readManagerRequestDTO)
+    // }
+    
+
+    // UPDATE - 점주신청서 승인 처리 (관리자 전용)
     @Put('/:request_id')
     @Roles(UserRole.ADMIN)
-    async updateManagerRequest(@Param('request_id') request_id: number, @Body() updateManagerRequestDTO: UpdateManagerRequestDTO): Promise<ApiResponseDTO<void>> {
+    async updateManagerRequest(
+        @Param('request_id') request_id: number,
+        @Body() updateManagerRequestDTO: UpdateManagerRequestDTO
+    ): Promise<ApiResponseDTO<void>> {
         await this.managerRequestsService.updateManagerRequest(request_id, updateManagerRequestDTO)
 
         return new ApiResponseDTO(true, HttpStatus.OK, "Request Updated Successfully")
     }
 
-    // DELETE
-    // 가게 신청서 삭제 
+    // DELETE - 점주 신청서 삭제 
     @Delete('/:request_id')
     @Roles(UserRole.MANAGER, UserRole.ADMIN)
     async deleteManagerRequest(
