+++ src/reviews/DTO/create-review.dto.ts
@@ -3,9 +3,5 @@ import { IsNotEmpty, IsNumber, IsString } from "class-validator"
 export class CreateReviewDTO {
     @IsNotEmpty()
     @IsString()
-    user_id: number
-
-    @IsNotEmpty()
-    @IsString()
     content: string
 }
