+++ src/reviews/reviews.controller.ts
@@ -1,6 +1,6 @@
 import { ApiResponseDTO } from 'src/common/api-reponse-dto/api-response.dto'
 import { ReviewsService } from './reviews.service'
-import { Body, Controller, Delete, ForbiddenException, Get, HttpStatus, Param, Patch, Put, Req, UseGuards } from '@nestjs/common'
+import { Body, Controller, Delete, ForbiddenException, Get, HttpStatus, Param, Patch, Post, Put, Req, UseGuards } from '@nestjs/common'
 import { Review } from './entites/review.entity'
 import { UpdateReviewDTO } from './DTO/update-review.dto'
 import { ReadReviewDTO } from './DTO/read-review.dto'
@@ -9,35 +9,65 @@ import { AuthGuard } from '@nestjs/passport'
 import { UserRole } from 'src/users/entities/user-role.enum'
 import { Roles } from 'src/common/custom-decorators/roles.decorator'
 import { AuthenticatedRequest } from 'src/auth/interfaces/authenticated-request.interface'
+import { CreateReviewDTO } from './DTO/create-review.dto'
 
-@Controller('api/reviews')
-@UseGuards(AuthGuard('jwt'), RolesGuard) // JWT인증, roles guard 적용
+@Controller('api/stores/:store_id/reviews')
+@UseGuards(AuthGuard('jwt'), RolesGuard)
 export class ReviewsController {
-
-    // 생성자 정의
     constructor(private reviewsService: ReviewsService) { }
 
-    // READ[1] - 모든 리뷰 조회 -> 안쓸듯
-    // 미구현: logger
-    @Get('/')
-    async readAllReviews(): Promise<ApiResponseDTO<ReadReviewDTO[]>> {
-        const reviews: Review[] = await this.reviewsService.readAllReviews()
-        const readReviewDTO = reviews.map(review => new ReadReviewDTO(review))
+    // CREATE - 새로운 리뷰 생성
+    @Post('/')
+    @Roles(UserRole.USER)
+    async createReview(
+        @Param('store_id') store_id: number,
+        @Req() req: AuthenticatedRequest,
+        @Body() createReviewDTO: CreateReviewDTO
+    ): Promise<ApiResponseDTO<void>> {
+        const user_id = req.user.user_id
 
-        return new ApiResponseDTO(true, HttpStatus.OK, 'Reviews Retrieved Successfully', readReviewDTO)
+        await this.reviewsService.createReview(store_id, user_id, createReviewDTO)
+
+        return new ApiResponseDTO(true, HttpStatus.CREATED, 'Review Created Successfully!')
+    }
+
+    // READ[1] 가게 리뷰 조회
+    @Get('/')
+    async readStoreReviews(
+        @Param('store_id') id: number
+    ): Promise<ApiResponseDTO<ReadReviewDTO[]>> {
+        const foundReviews = await this.reviewsService.readReviewsByStore(id)
+        const readReviewDTOs = foundReviews.map(review => new ReadReviewDTO(review))
+
+        return new ApiResponseDTO(true, HttpStatus.OK, "Store Reviews Retrieved Successfully", readReviewDTOs)
     }
 
     // READ[2] - 특정 리뷰 조회 -> 안쓸듯
-    // 미구현: looger
     @Get('/:review_id')
-    async readReviewById(@Param('review_id') review_id: number): Promise<ApiResponseDTO<Review>> {
+    async readReviewById(
+        @Param('review_id') review_id: number
+    ): Promise<ApiResponseDTO<Review>> {
         const foundReview: Review = await this.reviewsService.readReviewByReviewId(review_id)
         
         return new ApiResponseDTO(true, HttpStatus.OK, 'Successfully Retrieved Review!', foundReview)
     }
 
+    // READ - 나의 리뷰 조회
+    @Get('/my-reviews')
+    @UseGuards(AuthGuard('jwt'), RolesGuard)
+    @Roles(UserRole.USER)
+    async readMyReviews(
+        @Req() req: AuthenticatedRequest
+    ): Promise<ApiResponseDTO<ReadReviewDTO[]>> {
+        const user_id = req.user.user_id
+
+        const foundReviews = await this.reviewsService.readReviewsByUser(user_id)
+        const readReviewDTOs = foundReviews.map((review) => new ReadReviewDTO(review))
+
+        return new ApiResponseDTO(true, HttpStatus.OK, 'My Reviews Retrieved Successfully', readReviewDTOs)
+    }
+
     // UPDATE[1] - 본인 리뷰 수정
-    // 미구현: logger
     @Put('/:review_id')
     @Roles(UserRole.USER)
     async updateReviewByReviewId(
@@ -57,22 +87,16 @@ export class ReviewsController {
     }
 
     // UPDATE[2] - 리뷰 도움됐어요 반응
-    // 미구현: logger
-    /**
-     * 비고
-     * 한번만 누를 수 있게, 취소는 불가능
-     * 취소 되게하려면 복잡해지기 때문에 일단 이렇게
-     */
     @Patch('/:review_id/helpful')
     @Roles(UserRole.USER)
     async markHelpful(
-        @Param("review_id") review_id: number): Promise<ApiResponseDTO<void>> {
+        @Param("review_id") review_id: number
+    ): Promise<ApiResponseDTO<void>> {
         await this.reviewsService.markHelpful(review_id)
         return new ApiResponseDTO(true, HttpStatus.NO_CONTENT, 'Reaction Applied Successfully!')
     }
 
     // DELETE - 본인 리뷰 삭제
-    // 미구현: logger
     @Delete('/:review_id')
     @Roles(UserRole.USER, UserRole.ADMIN)
     async deleteReviewByReviewId(
@@ -81,7 +105,6 @@ export class ReviewsController {
     ): Promise<ApiResponseDTO<void>> {
         const foundReview = await this.reviewsService.readReviewByReviewId(review_id)
 
-        // USER는 본인 리뷰만 삭제 가능
         if (req.user.user_id !== foundReview.user.user_id) {
             throw new ForbiddenException('You Can Only Delete Your Own Review.')
         }
