+++ src/stores/stores.module.ts
@@ -1,20 +1,14 @@
-import { forwardRef, Module } from '@nestjs/common'
+import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { StoresService } from './stores.service'
 import { StoresController } from './stores.controller'
 import { Store } from './entities/store.entity'
-import { ReviewsModule } from 'src/reviews/reviews.module'
 import { UsersModule } from 'src/users/users.module'
-import { Event } from './entities/event.entity'
-import { Menu } from './entities/menu.entity'
 
 @Module({
     imports: [
         TypeOrmModule.forFeature([Store]),
-        TypeOrmModule.forFeature([Event]),
-        TypeOrmModule.forFeature([Menu]),
-        forwardRef(() => ReviewsModule),
-        forwardRef(() => UsersModule),
+        UsersModule,
     ],
     providers: [StoresService],
     controllers: [StoresController],
