+++ src/stores/stores.service.ts
@@ -1,36 +1,21 @@
 import { UsersService } from 'src/users/users.service'
 import { Injectable, NotFoundException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
-import { Store } from './entities/store.entity'
-import { LessThanOrEqual, MoreThanOrEqual, Repository } from 'typeorm'
+import { Repository } from 'typeorm'
 import { CreateStoreDTO } from './DTO/create-store.dto'
 import { StoreCategory } from './entities/store-category.enum'
 import { UpdateStoreDetailDTO } from './DTO/update-store-detail.dto'
-import { CreateEventDTO } from 'src/stores/DTO/create-event.dto'
-import { Event } from './entities/event.entity'
-import { UpdateEventDTO } from './DTO/update-event.dto'
-import { Menu } from './entities/menu.entity'
-import { CreateMenuDTO } from './DTO/create-menu.dto'
-import { UpdateMenuDTO } from './DTO/update-menu.dto'
+import { Store } from './entities/store.entity'
 
 @Injectable()
 export class StoresService {
-
-    // init
     constructor(
         @InjectRepository(Store)
         private storesRepository: Repository<Store>,
-        @InjectRepository(Event)
-        private eventRepository: Repository<Event>,
-        @InjectRepository(Menu)
-        private menusRepository: Repository<Menu>,
         private usersService: UsersService,
     ) { }
 
-    // ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ스토어 관련 기능 시작ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ
-
-    // CREATE
-    // 새로운 가게 생성하기
+    // CREATE - 새로운 가게 생성하기
     async createStore(user_id: number, createStoreDTO: CreateStoreDTO): Promise<Store> {
         const { store_name, category, address, latitude, longitude, contact_number, description } = createStoreDTO
 
@@ -63,8 +48,7 @@ export class StoresService {
         return foundStores
     }
 
-    // READ
-    // 모든 가게 조회
+    // READ - 모든 가게 조회
     async readAllStores(): Promise<Store[]> {
         const foundStores = await this.storesRepository.find()
 
@@ -92,8 +76,7 @@ export class StoresService {
         return foundStores
     }
 
-    // UPDATE
-    // 가게 매니저 속성 수정 (관리자 전용)
+    // UPDATE - 가게 매니저 속성 수정 (관리자 전용)
     async updateStoreManager(store_id: number, user_id: number): Promise<void> {
         const foundUser = await this.usersService.readUserById(user_id)
 
@@ -122,193 +105,10 @@ export class StoresService {
         await this.storesRepository.update(store_id, { public: true })
     }
 
-    // DELETE
-    // 가게 삭제하기
+    // DELETE - 가게 삭제하기
     async deleteStore(store_id: number): Promise<void> {
         const foundStore = await this.readStoreById(store_id)
 
         await this.storesRepository.remove(foundStore)
-    }
-
-
-    // ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ이벤트 관련 기능 시작ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ
-
-    // CREATE - 이벤트 생성
-    // 미구현: logger, 에러 처리
-    // 비고: 임시 시간값 사용
-    async createEvent(store_id: number, createEventDTO: CreateEventDTO): Promise<Event> {
-        const { title, description, start_date, end_date } = createEventDTO
-
-        // 가게 객체 가져오기
-        const store = await this.readStoreById(store_id)
-        if (!store) {
-            throw new NotFoundException(`Store with ID ${store_id} not found`)
-        }
-
-        // DTO에서 받은 값을 `Date` 객체로 변환, end_date가 명시되지 않았다면 7일 뒤로 설정
-        const startDate = start_date ? new Date(start_date) : new Date()
-        const endDate = end_date ? new Date(end_date) : new Date(startDate)
-        if (!end_date) {
-            endDate.setDate(endDate.getDate() + 7)
-        }
-
-        const newEvent: Event = this.eventRepository.create({
-            store,
-            title,
-            description,
-            start_date: startDate,
-            end_date: endDate,
-        })
-        const createdEvent = await this.eventRepository.save(newEvent)
-
-        return createdEvent
-    }
-
-    // READ - 최근 등록 이벤트 조회 (status: ONGOING 이벤트)
-    // 미구현: logger, 에러 처리
-    async readRecentEventByStore(store_id: number): Promise<Event> {
-        const now = new Date
-
-        const foundEvent = await this.eventRepository.findOne({
-            where: {
-                store: { store_id },
-                start_date: LessThanOrEqual(now),
-                end_date: MoreThanOrEqual(now)
-            },
-            order: { created_at: 'DESC' },
-            relations: ['store']
-        })
-        if (!foundEvent) {
-            throw new NotFoundException(`Cannot Find Event In This Store`)
-        }
-
-        return foundEvent
-    }
-
-    // READ - 해당 가게의 모든 이벤트 조회 (생성일 기준 정렬)
-    // 미구현: logger, 에러 처리
-    async readAllEventsByStore(store_id: number): Promise<Event[]> {
-        const foundEvents = await this.eventRepository.find({
-            where: { store: { store_id } },
-            order: { created_at: 'DESC' },
-        })
-        if (!foundEvents) {
-            throw new NotFoundException(`Cannot Find Events`)
-        }
-
-        return foundEvents
-    }
-
-    // READ - 특정 이벤트 상세 조회
-    // 미구현: logger, 에러 처리
-    async readEventById(event_id: number): Promise<Event> {
-        const foundEvent = await this.eventRepository.findOne({
-            where: { event_id },
-            relations: ['store.user']
-        })
-        if (!foundEvent) {
-            throw new NotFoundException(`Cannot Find Event by Id ${event_id}`)
-        }
-
-        return foundEvent
-    }
-
-    // UPDATE - by event_id
-    // 미구현: logger, 에러 처리
-    async updateEventById(event_id: number, updateEventDTO: UpdateEventDTO) {
-        const foundEvent = await this.readEventById(event_id)
-
-        const { title, description, start_date, end_date, is_canceled } = updateEventDTO
-
-        foundEvent.title = title
-        foundEvent.description = description
-        foundEvent.start_date = start_date
-        foundEvent.end_date = end_date
-        foundEvent.is_canceled = is_canceled
-
-        await this.eventRepository.save(foundEvent)
-    }
-
-    // DELETE
-    // 미구현: logger, 에러 처리
-    async deleteEventById(event_id: number) {
-        const foundEvent = await this.readEventById(event_id)
-
-        await this.eventRepository.remove(foundEvent)
-    }
-
-
-    // ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ메뉴 관련 기능 시작ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ
-
-    // CREATE - 새로운 메뉴 등록
-    // 미구현: logger, 에러 처리
-    async createMenu(store_id: number, createMenuDTO: CreateMenuDTO): Promise<Menu> {
-        const { menu_name, price, description } = createMenuDTO
-
-        // 가게 객체 가져오기
-        const store = await this.readStoreById(store_id)
-        if (!store) {
-            throw new NotFoundException(`Store with ID ${store_id} not found`)
-        }
-
-        const newMenu: Menu = this.menusRepository.create({
-            store,
-            menu_name,
-            price,
-            description,
-        })
-
-        const createdMenu = await this.menusRepository.save(newMenu)
-
-        return createdMenu
-    }
-
-    // READ[1] - 해당 가게 모든 메뉴 조회
-    // 미구현: logger, 에러 처리
-    async readMenusByStore(store_id: number): Promise<Menu[]> {
-        const foundMenus = await this.menusRepository.find({
-            where: { store: { store_id } }
-        })
-        if (!foundMenus) {
-            throw new NotFoundException(`Cannot Find Menus`)
-        }
-
-        return foundMenus
-    }
-
-    // READ[2] - 특정 메뉴 상세 조회
-    // 미구현: logger, 에러 처리
-    async readMenuById(menu_id: number): Promise<Menu> {
-        const foundMenu = await this.menusRepository.findOne({
-            where: { menu_id },
-            relations: ['store.user']
-        })
-        if (!foundMenu) {
-            throw new NotFoundException(`Cannot Find Menu By Id ${menu_id}`)
-        }
-        return foundMenu
-    }
-
-    // UPDATE - by menu_id
-    // 미구현: logger, 에러 처리
-    async updateMenuById(menu_id: number, updateMenuDTO: UpdateMenuDTO) {
-        const foundMenu = await this.readMenuById(menu_id)
-
-        const { menu_name, description, price, manager_container } = updateMenuDTO
-        foundMenu.menu_name = menu_name
-        foundMenu.description = description
-        foundMenu.price = price
-        foundMenu.manager_container = manager_container
-
-        await this.menusRepository.save(foundMenu)
-    }
-
-    // DELETE
-    // 미구현: logger, 에러 처리
-    async deleteMenuById(menu_id: number) {
-        const foundMenu = await this.readMenuById(menu_id)
-        if (foundMenu) {
-            await this.menusRepository.remove(foundMenu)
-        }
     }
 }
